I created a branch called instructPartThree. This branch completes the requirement called Part 3: Event Listeners
In the branch instructPartThree, I modified the files index.html, main.css and main.js. In index.html, i added all of the buttons with the proper classes and created a div around those buttons. I also changed the div id on line 18 from petsToHTML to app. On main.css, I changed the id on line 5 to app. I also added classes and elements of the filtering buttons to look better, as well as place more margin on the bottoms of the cards to display better. Lastly, for main.js, this file received the most changes. Each piece of JS code has NEW NOTES and OLD NOTES. Focus on the new notes to receive a more efficient explanation. AVOID OLD NOTES