I created a branch called instructPartFour. This branch completes the Requirement Part Four on instructions.md
In the branch instructPartFour, I modified index.html, main.css and main.js. In index.html, I created a form that corresponds with the functions on JS. In main.css, I added the width of the card onto the class of .card and deleted it from main.js (line 288). In main.js, I created a function that targets the form and collects the values from it to push onto the array of existing pets. Lastly in main.js, I created an event listener to listen for when the submit button is clicked(submitted) and to pass the function that was created. NEW NOTES explains the process of each component for clarification. AVOID OLD NOTES.